refactor: rename control_packet.h to packet_control.h
Closed #50
